Rearrange code, extract documentation collector
